src/requirements.txt
SQLAlchemy
psycopg2-binary  
python-dotenv
tabulate   